rename pillar color utility file
