+++ client-v2/src/shared/util/getPillarColor.ts
